move the vat files for the handoff tests to test/swingsetTests/scaffolding/
