+++ test/swingsetTests/scaffolding/handoff/bootstrap.js
@@ -2,8 +2,8 @@
 
 import harden from '@agoric/harden';
 
-import { makeCorkboard } from '../../more/handoff/corkboard';
-import { makeHandoffService } from '../../more/handoff/handoff';
+import { makeCorkboard } from '../../../../more/handoff/corkboard';
+import { makeHandoffService } from '../../../../more/handoff/handoff';
 
 function build(E, log) {
   function testCorkboardStorage() {
