+++ test/swingsetTests/scaffolding/handoff/vat-bob.js
